Add key prop to TabContent
So that the state is reset when switching to another tab. Because of the React diffing algorithm the state would be kept without the key prop. The component would stay at the same place, because only the tab content changes so then React doesn't destroy the component instance, but reuses it.